+++ src/utils.cpp
@@ -193,11 +193,11 @@ KNode<data_type> *convert_to_knodes(data_type *tree, int size, int n_components,
                                 n_components, nullptr, nullptr, false);
 }
 
-int sort_and_split(DataPoint *array, int size, int axis) {
+size_t sort_and_split(DataPoint *array, int size, int axis) {
   // the second part of median_idx is needed to unbalance the split towards
   // the left region (which is the one which may parallelize with the highest
   // probability).
-  int median_idx = size / 2 - 1 * ((size + 1) % 2);
+  size_t median_idx = size / 2 - 1 * ((size + 1) % 2);
   std::nth_element(array, array + median_idx, array + size,
                    DataPointCompare(axis));
   // if size is 2 we want to return the first element (the smallest one),
@@ -205,13 +205,14 @@ int sort_and_split(DataPoint *array, int size, int axis) {
   return median_idx;
 }
 
-int sort_and_split(std::vector<DataPoint>::iterator first_data_point,
-                   std::vector<DataPoint>::iterator last_data_point, int axis) {
-  int size = std::distance(first_data_point, last_data_point);
+size_t sort_and_split(std::vector<DataPoint>::iterator first_data_point,
+                      std::vector<DataPoint>::iterator last_data_point,
+                      int axis) {
+  size_t size = std::distance(first_data_point, last_data_point);
   // the second part of median_idx is needed to unbalance the split towards
   // the left region (which is the one which may parallelize with the highest
   // probability).
-  int median_idx = size / 2 - 1 * ((size + 1) % 2);
+  size_t median_idx = size / 2 - 1 * ((size + 1) % 2);
   std::nth_element(first_data_point, first_data_point + median_idx,
                    last_data_point, DataPointCompare(axis));
   // if size is 2 we want to return the first element (the smallest one),
